reduced test coverage to 73% but made sure every package is tested
