+++ .github/workflows/tests.yaml
@@ -19,10 +19,10 @@ jobs:
       run: |
         export COVERAGE=$(go tool cover -func coverage.out | grep total | awk '{print substr($3, 1, length($3)-1)}')
         echo "code coverage is at ${COVERAGE}"
-        if [ 1 -eq "$(echo "${COVERAGE} > 74.0" | bc)" ]; then \
-          echo "all good... coverage is above 74.0%"; 
+        if [ 1 -eq "$(echo "${COVERAGE} > 73.0" | bc)" ]; then \
+          echo "all good... coverage is above 73.0%"; 
         else \
-          echo "not good... coverage is not above 74.0%";
+          echo "not good... coverage is not above 73.0%";
           exit 1
         fi
     - name: Upload coverage to Codecov
