+++ base/test_helpers.go
@@ -13,3 +13,20 @@ func SetupWithMock(t *testing.T) {
 		httpmock.NewStringResponder(200, `[{"id": 1, "name": "My Great Thing"}]`))
 	t.Cleanup(httpmock.DeactivateAndReset)
 }
+
+type mockDriver struct {
+	*Experiment
+}
+
+func (m *mockDriver) ReadResult() (*ExperimentResult, error) {
+	return m.Experiment.Result, nil
+}
+
+func (m *mockDriver) WriteResult(r *ExperimentResult) error {
+	m.Experiment.Result = r
+	return nil
+}
+
+func (m *mockDriver) ReadSpec() (ExperimentSpec, error) {
+	return m.Experiment.Tasks, nil
+}
